Moves scheduled background threads into a separate background process
