+++ requirements.txt
@@ -9,6 +9,7 @@ email_validator==1.0.2
 phonenumbers==8.8.1
 pockets>=0.6.1
 residue>=0.2.4
+schedule>=0.5.0
 XlsxWriter==1.0.2
 uszipcode==0.1.3
 geopy==1.11.0
